Create TransposeMelt_subpubnet
This file is the sub pub author network of US patent authors. It originated from a df that is a merge of the metadata for US patent authors from the ReferencePatent table inner joined on name string matches to the disamb_authors table, inner joined to the Reference table. Also, Annotation table data was later merged into this data frame to provide additional information about the submission (public release of the data submission) date, organism type, among other data. The authors were best formatted (all authors, comma separated in a single column) in this df for subs and pubs, so for a test (in anticipation of the authors patent network which we can rbind to this initial file) we will work with this subset. It gives us the forknowledge that there will be some connectivity in the colaboration network of sub, pub, patent, so we do not waster time and effort tot find that a graph produced has no overlap, that is, that there are no authors who are on patents who are also authors of subs and pubs, that is, the giant component represents >75% of the network (ideally - this is an arbitrary number I would hypothesize that we see, BTW). So, the type of reference (sub or pub) and year are in this initial dataset. 

In this file,we take this dataset of merged US patent, sub, pub, and format it properly for the sub pub authors and type of Ref. To do this, we create author pairs first cleaning the author column (gsub the "and "s) and then parsing the authors field by ".,". After cleaning and parsing, we prepare for formatting for the network analysis and visualization. To do this, we must first isolate the authors and their associated data: the years go into their own 1-column vector and ref type go into their own 1-column vector, preserving the order in the current df to make sure the year and Ref type are correctly associated with the authors on that Reference (i.e., in that row). Then, get just the authors into a df (without year and type) and we calculate every combination of a size=2 (i.e. pairs) and transpose them into a df of these pairs, so every row is an author, author dyad. There were some solo authors, so we had to remove these because there is no place for a solo author in the collaboration network....he will show up only if she is a collaborator (This solo-authorship is something to explore, especially the ratio of solo:collab; that is, the number of solo-authored References to the number of collaboratively produced references. Very interesting, because it perhaps comments on knowledge production processes of scientists, and their perhaps yearly rhythms of pub, sub, and patent work contributions). 
Then, we add to this data frame (this is just the first row, mind you, before we write the function to iterate through) the first row year and type. Then, we write the iteration function and let it run.